requirements.txt
# Core Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# HTTP Requests
requests>=2.31.0
httpx>=0.27.0

# Database
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Data Models & Validation
pydantic>=2.5.0
pydantic-settings>=2.0.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Configuration
python-dotenv>=1.0.0

# File Upload Support
python-multipart>=0.0.6

# Type Hints (for older Python versions)
typing-extensions>=4.8.0

# Date/Time utilities
python-dateutil>=2.8.0
